requirements.txt
asgiref==3.8.1
chardet==5.2.0
distlib==0.3.9
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.16.1
pbr==6.1.0
pillow==11.0.0
platformdirs==4.3.6
PyJWT==2.9.0
PyPDF2==3.0.1
reportlab==4.2.5
sqlparse==0.5.1
stevedore==5.3.0
typing_extensions==4.12.2
virtualenv==20.27.1
virtualenv-clone==0.5.7
virtualenvwrapper==6.1.1
